+++ website/home/models/settings.py
@@ -1,0 +1,39 @@
+from wagtail.contrib.settings.models import (
+    BaseGenericSetting,
+    register_setting,
+)
+from wagtail.fields import RichTextField
+from wagtail.models import Page
+from wagtail.admin.panels import FieldPanel
+from django.db import models
+from django.conf import settings
+from django import forms
+
+
+@register_setting
+class Footer(BaseGenericSetting):
+    technicalQuestions = RichTextField(
+        blank=True, help_text="Content for technical questions."
+    )
+    otherQuestions = RichTextField(blank=True, help_text="Content for other questions.")
+
+    content_panels = Page.content_panels + [
+        FieldPanel("technicalQuestions"),
+        FieldPanel("otherQuestions"),
+    ]
+
+
+@register_setting
+class GoogleAnalyticsSettings(BaseGenericSetting):
+    tracking_id = models.CharField(
+        max_length=20,
+        help_text="Google Analytics Measurement ID (e.g., G-1234567890)",
+        default=settings.GOOGLE_ANALYTICS_ID,
+    )
+
+    panels = [
+        FieldPanel(
+            "tracking_id",
+            widget=forms.TextInput(attrs={"value": settings.GOOGLE_ANALYTICS_ID}),
+        ),
+    ]
